Fix: Codacy error in educators evaluation
