+++ src/screens/educators/EducatorKidsEvaluationYearly.js
@@ -1,125 +1,42 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import axios from 'axios';
 import LayoutProfiles from '../../components/LayoutProfiles';
 import StudentEvaluation from '../../components/StudentEvaluation';
 import '../../styles/styles.css';
-import { toast,ToastContainer } from 'react-toastify';
+import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import EducatorEvaluationCommon from '../../components/Evaluation';
 const API_ENDPOINT = process.env.REACT_APP_API_ENDPOINT
 
 function EducatorKidsEvaluationYearly() {
-  const [userId,setUserId] = useState(null);
-  const [educatorId,setEducatorId] = useState(null);
-  const [kids,setKids] = useState(null);
-  const [students,setStudents] = useState(null);
-  const [evaluation,setEvaluation] = useState({});
-  const [selectedStudent,setSelectedStudent] = useState(null);
-  const [showEditModal,setShowEditModal] = useState(false);
-  const [showInfoModal,setShowInfoModal] = useState(false);
-  const [comment, setComment] = useState("");
-  const [grade, setGrade] = useState("");
-  const [selectedDate, setSelectedDate] = useState(null);
-  const [studentEvaluations, setStudentEvaluations] = useState([]);
-  const [email, setEmail] = useState("");
-  const [phone, setPhone] = useState("");
-  const handleEvaluationChange = (id) => (event) => {
-    setEvaluation((prevEvaluation) => ({
-      ...prevEvaluation,
-      [id]: event.target.value,
-    }));
-  };
-
+  const {
+    userId, setUserId,
+    educatorId, setEducatorId,
+    kids, setKids,
+    students, setStudents,
+    setEvaluation,
+    selectedStudent,
+    showEditModal,
+    showInfoModal, 
+    comment,
+    grade, 
+    selectedDate, 
+    email, setEmail,
+    handleEvaluationChange,
+    handleEdit,
+    handleInfo,
+    handleCloseModal,
+    handleCloseInfoModal,
+    handleCommentChange,
+    handleGradeChange,
+    handleSubmit,
+    handleDateChange,
+    getStudentEvaluation,
+    phone,setPhone,
+    setStudentEvaluations
+  } = EducatorEvaluationCommon();
+  
  
-  const handleEdit = (id) => {
-    const student = students.find(student => student.id === id);
-    setSelectedStudent(student);
-    setShowEditModal(true);
-  };
-  const handleInfo = (id) => {
-    const student = students.find(student => student.id === id);
-    setSelectedStudent(student);
-    setShowInfoModal(true);
-  };
-
-  const handleCloseModal = () => {
-    setShowEditModal(false);
-  };
-  const handleCloseInfoModal = () => {
-    setShowInfoModal(false);
-  };
-  const handleCommentChange = (event) => {
-    setComment(event.target.value);
-  };
-
-  const handleGradeChange = (event) => {
-    setGrade(event.target.value);
-  };
-
-  const handleSubmit = async () => {
-    try {
-      console.log('current date',selectedDate)
-   
-      const studentId = selectedStudent.id;
-      const { data: evaluations } = await axios.get(`${API_ENDPOINT}student-evaluation/`);
-      console.log('Fetched evaluations:', evaluations);
-  
-      // Check if an evaluation already exists for the current year
-      const currentYear = new Date().getFullYear();
-      const studentEvaluation = evaluations.find(evaluation => 
-        evaluation.student === parseInt(studentId) && 
-        evaluation.evaluation_type === 1 &&
-        new Date(evaluation.date).getFullYear() === currentYear
-      );
-      console.log('Found student evaluation:', studentEvaluation);
-  
-      if (studentEvaluation) {
-        console.log('Updating evaluation with grade:', grade, 'and comment:', comment);
-        const updateResponse = await axios.put(`${API_ENDPOINT}student-evaluation/${studentEvaluation.id}/`, {
-          ...studentEvaluation,
-          grade: parseInt(grade, 10), // Update the grade
-          comment: comment, // Update the comment
-        });
-        console.log('Update response:', updateResponse);
-        if (updateResponse.status === 200) {
-          console.log('Evaluation updated successfully');
-          toast.success('Evaluación realizada de manera correcta');
-          setStudentEvaluations(prevEvaluations => prevEvaluations.map(evaluation => 
-            evaluation.id === studentEvaluation.id ? updateResponse.data : evaluation
-          ));
-          handleCloseModal();
-        } else {
-          console.log('Failed to update evaluation');
-          toast.error('Error al evaluar al estudiante');
-        }
-      } else {
-        console.log('Creating new evaluation with grade:', grade, 'and comment:', comment);
-        const createResponse = await axios.post(`${API_ENDPOINT}student-evaluation/`, {
-          grade: parseInt(grade, 10), // Parse grade to integer
-          date: selectedDate, // Use current date
-          comment: comment,
-          student: parseInt(studentId),
-          evaluation_type: 1,
-        });
-        console.log('Create response:', createResponse);
-      
-        if (createResponse.status === 201) {
-          console.log('Evaluation created successfully');
-          toast.success('Evaluación realizada de manera correcta');
-          setStudentEvaluations(prevEvaluations => [...prevEvaluations, createResponse.data]);
-          handleCloseModal();
-        } else {
-          console.log('Failed to create evaluation');
-          toast.error('Error al evaluar al estudiante');
-        }
-      }
-      } catch (error) {
-        console.error('An error occurred:', error);
-        toast.error('Todos los campos son obligatorios');
-      }
-  };
-  const handleDateChange = (event) => {
-    setSelectedDate(event.target.value);
-  };
 
   useEffect (()=>{
     const id = localStorage.getItem('userId');
@@ -172,27 +89,9 @@ function EducatorKidsEvaluationYearly() {
         });
     }
   }, [kids]);
- const getStudentEvaluation = (studentId) => {
-  const filteredEvaluations = studentEvaluations.filter(evaluation => 
-    evaluation.student === parseInt(studentId) && 
-    (evaluation.evaluation_type === 1 )
-  );
 
-  if (filteredEvaluations.length === 0) {
-    return 'Sin evaluar';
-  }
-  
-  filteredEvaluations.sort((a, b) => new Date(b.date) - new Date(a.date));
 
-  // Check if the most recent evaluation is from the current year
-  const currentYear = new Date().getFullYear();
-  if (new Date(filteredEvaluations[0].date).getFullYear() === currentYear) {
-    // Return the grade of the most recent evaluation
-    return filteredEvaluations[0].grade;
-  } else {
-    return 'Sin evaluar';
-  }
-};
+
   useEffect(() => {
     if (selectedStudent) {
       axios.get(`${API_ENDPOINT}user/`)
@@ -213,7 +112,7 @@ function EducatorKidsEvaluationYearly() {
        <ToastContainer />
        <StudentEvaluation 
         students={students}
-        evaluationType={2}
+        evaluationType={1}
         grade={grade}
         handleGradeChange={handleGradeChange}
         comment={comment}
